moved code as suggested by Tim
