+++ dataservices/tests/test_helpers.py
@@ -19,6 +19,21 @@ from sqlalchemy.future.engine import Engine
 
 from dataservices import helpers, models
 from dataservices.management.commands import helpers as dmch
+from dataservices.management.commands.import_dbt_investment_opportunities import (
+    get_investment_opportunities_batch,
+    get_investment_opportunities_data_table,
+    save_investment_opportunities_data,
+)
+from dataservices.management.commands.import_dbt_sectors import (
+    get_dbtsector_postgres_table,
+    get_dbtsector_table_batch,
+    save_dbt_sectors_data,
+)
+from dataservices.management.commands.import_sectors_gva_value_bands import (
+    get_sectors_gva_value_bands_batch,
+    get_sectors_gva_value_bands_table,
+    save_sectors_gva_value_bands_data,
+)
 from dataservices.tests import factories, utils
 
 
@@ -333,14 +348,14 @@ def test_unzip_s3_gzip_file_eof(mock_decompressobj):
 @mock.patch.object(Engine, 'connect')
 def test_save_dbtsector_data(mock_connection, mock_ingest, dbtsector_data):
     mock_connection.return_value.__enter__.return_value = mock.MagicMock()
-    dmch.save_dbt_sectors_data(data=dbtsector_data)
+    save_dbt_sectors_data(data=dbtsector_data)
     assert mock_ingest.call_count == 1
 
 
 @pytest.mark.django_db
 def test_get_dbtsector_table_batch(dbtsector_data):
     metadata = sa.MetaData()
-    ret = dmch.get_dbtsector_table_batch(dbtsector_data, dmch.get_dbtsector_postgres_table(metadata))
+    ret = get_dbtsector_table_batch(dbtsector_data, get_dbtsector_postgres_table(metadata))
     assert next(ret[2]) is not None
 
 
@@ -350,16 +365,14 @@ def test_get_dbtsector_table_batch(dbtsector_data):
 @mock.patch.object(Engine, 'connect')
 def test_save_sectors_gva_value_bands_data(mock_connection, mock_ingest, sectors_gva_value_bands_data):
     mock_connection.return_value.__enter__.return_value = mock.MagicMock()
-    dmch.save_sectors_gva_value_bands_data(data=sectors_gva_value_bands_data)
+    save_sectors_gva_value_bands_data(data=sectors_gva_value_bands_data)
     assert mock_ingest.call_count == 1
 
 
 @pytest.mark.django_db
 def test_get_sectors_gva_value_bands_batch(sectors_gva_value_bands_data):
     metadata = sa.MetaData()
-    ret = dmch.get_sectors_gva_value_bands_batch(
-        sectors_gva_value_bands_data, dmch.get_sectors_gva_value_bands_table(metadata)
-    )
+    ret = get_sectors_gva_value_bands_batch(sectors_gva_value_bands_data, get_sectors_gva_value_bands_table(metadata))
     assert next(ret[2]) is not None
 
 
@@ -369,14 +382,14 @@ def test_get_sectors_gva_value_bands_batch(sectors_gva_value_bands_data):
 @mock.patch.object(Engine, 'connect')
 def test_save_investment_opportunities_data(mock_connection, mock_ingest, investment_opportunities_data):
     mock_connection.return_value.__enter__.return_value = mock.MagicMock()
-    dmch.save_investment_opportunities_data(data=investment_opportunities_data)
+    save_investment_opportunities_data(data=investment_opportunities_data)
     assert mock_ingest.call_count == 1
 
 
 @pytest.mark.django_db
 def test_get_investment_opportunities_batch(investment_opportunities_data):
     metadata = sa.MetaData()
-    ret = dmch.get_investment_opportunities_batch(
-        investment_opportunities_data, dmch.get_investment_opportunities_data_table(metadata)
+    ret = get_investment_opportunities_batch(
+        investment_opportunities_data, get_investment_opportunities_data_table(metadata)
     )
     assert next(ret[2]) is not None
